requirements.txt
requests==2.31.0
PyQt5==5.15.9
python-dotenv==1.0.0
